requirements.txt
streamlit==1.44.0
google-genai==1.68.0
